requirements.txt
flask
pandas
numpy
scikit-learn
mrmrpy
openpyxl
xlsxwriter
gunicorn